SKALE-2266 Retrieve util methods in separate module
